Update Disable_Hyper_V_and_Device_Guard.ps1
Added a helper function to get Windows Feature status and convert to string, with graceful handling of undefined features.  

Modified logic of save_config function to make use of this helper.  

Impact of this is that featureState should now be persisted out to the config JSON file in a format / data type that is compatible with the load_config function, addressing the bug where settings would not revert correctly.